Feat(LOCAT-BE-37): merge class
