+++ src/main/java/com/locat/api/global/notification/PlatformEndpointService.java
@@ -1,5 +1,19 @@
 package com.locat.api.global.notification;
 
 public interface PlatformEndpointService {
+
+    /**
+     * 디바이스 토큰과 플랫폼을 기준으로 엔드포인트를 생성합니다.
+     * @param token 디바이스 토큰
+     * @param platform 플랫폼 - ios, android
+     * @return 생성된 디바이스별 엔드포인트 arn
+     */
     String createPlatformEndpoint(String token, String platform);
+
+    /**
+     * 엔드포인트를 AWS SNS Topic에 구독시킵니다.
+     * @param endpointArn 디바이스별 엔드포인트 arn
+     * @return 구독 arn
+     */
+    String subscribeEndpointToTopic(String endpointArn);
 }
